# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  api.SSETicketResponse:
    properties:
      expires_in:
        type: integer
      ticket:
        type: string
    type: object
  api.TokenResponse:
    properties:
      access_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/api.UserResponse'
    type: object
...
  api.UserResponse:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      org_id:
        type: integer
      org_name:
        type: string
      role:
        type: string
    type: object
...
  api.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  api.signupRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      org_name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
host: localhost:8011
info:
  contact:
    email: support@callified.ai
    name: Callified Support
  description: AI-powered outbound calling, WhatsApp CRM & receptionist platform.
  termsOfService: http://callified.ai/terms
  title: Callified Backend API
  version: "1.0"
paths:
...
  /api/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Sends a password-reset link to the given email. Returns the same
        message regardless of whether the email exists.
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Forgot password
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email + password. Returns a 30-day JWT.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: invalid credentials
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Login
      tags:
      - auth
  /api/auth/me:
    get:
      description: Returns the authenticated user's profile and org name.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Validates the reset token and sets a new password.
      parameters:
      - description: Reset token and new password
        in: body
        name: body
        required: true
        schema:
          properties:
            new_password:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Reset password
      tags:
      - auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account and organisation. Returns a JWT on success.
      parameters:
      - description: Signup payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.signupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: email already registered
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Sign up
      tags:
      - auth
...
  /api/sse/ticket:
    get:
      description: Issues a 60-second short-lived JWT for use as ?ticket= on SSE endpoints.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SSETicketResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mint SSE ticket
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: **Bearer {JWT token}**. Obtain a token from POST /api/auth/login
      or POST /api/auth/signup.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"